bark-bot/pyproject.toml
[project]
name = "bark-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "boto3>=1.42.59",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "firecrawl-py>=4.18.0",
    "google-api-python-client>=2.191.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "notion-client>=3.0.0",
    "openai>=2.24.0",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pygithub>=2.8.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "slack-sdk>=3.40.1",
    "sqlalchemy>=2.0.47",
    "tavily-python>=0.7.22",
    "uvicorn[standard]>=0.41.0",
]
